re-generated the bounds using cnn_no_softmax and preds_out_of_bound now runs without errors
previously the problem was that the mnist_ids of the images in the test dataloader (and so in the test_indices) did not exist within the results files, i.e. these images did not have bounds computed. I don't know how this happened but re-running verification_saved_models fixed it by generating the bounds for the correct images somehow.